requirements.txt
-e .
requests>=2.28
numpy>=1.24
pandas>=2.0
scipy>=1.10
matplotlib>=3.7
openmeteo-requests>=1.0
requests-cache>=1.0
retry-requests>=2.0
streamlit>=1.35
plotly>=5.22
mysql-connector-python>=8.0
python-dotenv>=1.0
pyyaml>=6.0
lxml>=4.9
